fixing large sets
Summary:
Sets with a large number of elements (>800000000) are not supported by IntOpenHashSet.
Chaning it to IntOpenHashBigSet.

ArrayLists have the same problem, but we'll postpone the fix untill we have a use case

https://issues.apache.org/jira/browse/GIRAPH-1028

Test Plan:
mvn clean install
see also a new test TestCollections (that needs 32G to run)

Reviewers: sergey.edunov, maja.kabiljo, ikabiljo

Reviewed By: ikabiljo

Differential Revision: https://reviews.facebook.net/D44859

+++ giraph-core/src/main/java/org/apache/giraph/types/ops/IntTypeOps.java
@@ -68,7 +68,7 @@ public enum IntTypeOps
   }
 
   @Override
-  public BasicSet<IntWritable> createOpenHashSet(int capacity) {
+  public BasicSet<IntWritable> createOpenHashSet(long capacity) {
     return new BasicIntOpenHashSet(capacity);
   }
